rag_chatbot/requirements.txt
python-dotenv>=1.0.1
langchain>=0.1.16
langchain-community>=0.0.32
langchain-openai>=0.1.6
langchain-google-genai>=2.0.0
langchain-chroma>=0.1.0
langchain-core>=0.1.0
chromadb
requests
fastapi
uvicorn
pydantic